third_party/WebKit/LayoutTests/bluetooth/getPrimaryService/device-disconnects-during-success.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
  return setBluetoothFakeAdapter('DisconnectingHeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}],
      optionalServices: [request_disconnection_service_uuid]
    }))
    .then(device => {
      return device.gatt.connect()
        .then(gattServer => gattServer.getPrimaryService(request_disconnection_service_uuid))
        .then(service => service.getCharacteristic(request_disconnection_characteristic_uuid))
        .then(requestDisconnection => {
          requestDisconnection.writeValue(new Uint8Array([0]));
          return assert_promise_rejects_with_message(
            device.gatt.getPrimaryService('heart_rate'),
            new DOMException('GATT Server disconnected while retrieving services.',
                             'NetworkError'));
        });
    });
}, 'Device disconnects during a getPrimaryService call that succeeds. Reject ' +
   'with NetworkError.');
</script>
